fix: last few ui updates
