+++ src/components/resume/finances/budgeting/getBudgetPieContent.ts
@@ -1,0 +1,70 @@
+import type { Theme } from "@mui/material/styles";
+import type {
+  ExpenseEntry,
+  ExpenseEntryColor,
+} from "../../../../jotai/finances-atom";
+import {
+  colorizeBreakdownPieData,
+  colorizeIncomeOverviewPieData,
+} from "./graphs/chartColors";
+import {
+  buildExpensePieData,
+  buildIncomeOverviewPieData,
+  buildPayrollPieData,
+  PAYROLL_CATEGORY_KEY,
+  type BudgetFlow,
+  type CategoryTotal,
+  type PiePoint,
+} from "./helpers";
+
+export interface BudgetPieContent {
+  title: string;
+  data: PiePoint[];
+}
+
+export const getBudgetPieContent = (
+  flow: BudgetFlow,
+  categories: CategoryTotal[],
+  selectedCategoryKey: string | null,
+  expenseEntries: ExpenseEntry[],
+  categoryColors: Partial<Record<string, ExpenseEntryColor>>,
+  theme: Theme,
+): BudgetPieContent => {
+  if (selectedCategoryKey === PAYROLL_CATEGORY_KEY) {
+    return {
+      title: "Payroll Breakdown",
+      data: colorizeBreakdownPieData(theme, "error", buildPayrollPieData(flow)),
+    };
+  }
+
+  if (selectedCategoryKey) {
+    const selectedCategory = categories.find(
+      ({ categoryKey }) => categoryKey === selectedCategoryKey,
+    );
+    const data = buildExpensePieData(
+      selectedCategoryKey,
+      expenseEntries,
+      flow.income,
+    );
+
+    return {
+      title: selectedCategory
+        ? `${selectedCategory.heading} Breakdown`
+        : "Category Breakdown",
+      data: colorizeBreakdownPieData(
+        theme,
+        selectedCategory?.color ?? categoryColors[selectedCategoryKey],
+        data,
+      ),
+    };
+  }
+
+  return {
+    title: "Income Overview",
+    data: colorizeIncomeOverviewPieData(
+      theme,
+      categories,
+      buildIncomeOverviewPieData(flow),
+    ),
+  };
+};
