+++ src/components/resume/finances/budgeting/graphs/BudgetSankeyGraph.test.tsx
@@ -2,16 +2,18 @@ import { clickSankeyNode } from "./tests/highchartsMocks";
 import { render, screen } from "@testing-library/react";
 import { createStore, Provider } from "jotai";
 import { describe, expect, it, vi } from "vitest";
-import themeAtom, { lightTheme } from "../../../../jotai/theme-atom";
+import themeAtom, { lightTheme } from "../../../../../jotai/theme-atom";
 import BudgetSankeyGraph from "./BudgetSankeyGraph";
 import {
   buildBudgetFlow,
   getLatestBudgetIncome,
+} from "../../../../../apis/budget";
+import {
   GROSS_INCOME_NODE,
   INCOME_NODE_LABELS,
   PAYROLL_CATEGORY_KEY,
-} from "./helpers";
-import { PAYROLL_NODE_LABEL } from "../../../../constants/payrollDeductions";
+} from "./chartData";
+import { PAYROLL_NODE_LABEL } from "../../../../../constants/payrollDeductions";
 
 const sampleFlow = buildBudgetFlow(
   getLatestBudgetIncome(100_000, 10_000, 0, 0),
@@ -45,7 +47,7 @@ describe("resume | finances | budgeting | BudgetSankeyGraph", () => {
       </Provider>,
     );
 
-    clickSankeyNode({ id: "HOUSING" });
+    clickSankeyNode({ id: "Housing" });
 
     expect(onCategorySelect).toHaveBeenCalledWith("housing");
   });
@@ -96,7 +98,7 @@ describe("resume | finances | budgeting | BudgetSankeyGraph", () => {
       </Provider>,
     );
 
-    clickSankeyNode({ id: "HOUSING", isNode: false });
+    clickSankeyNode({ id: "Housing", isNode: false });
 
     expect(onCategorySelect).not.toHaveBeenCalled();
   });
@@ -164,6 +166,7 @@ describe("resume | finances | budgeting | BudgetSankeyGraph", () => {
     );
 
     clickSankeyNode({ id: "" });
+    clickSankeyNode({ id: undefined as unknown as string });
 
     expect(onCategorySelect).not.toHaveBeenCalled();
   });
