+++ src/components/resume/finances/budgeting/graphs/chartColors.test.ts
@@ -8,9 +8,12 @@ import {
   getBudgetSankeyNodeColors,
   mixHexColors,
 } from "./chartColors";
-import { buildBudgetFlow, getLatestBudgetIncome } from "./helpers";
-import { FEDERAL_TAX_LABEL } from "../../../../constants/federalTaxBrackets";
-import { PAYROLL_NODE_LABEL } from "../../../../constants/payrollDeductions";
+import {
+  buildBudgetFlow,
+  getLatestBudgetIncome,
+} from "../../../../../apis/budget";
+import { FEDERAL_TAX_LABEL } from "../../../../../constants/federalTaxBrackets";
+import { PAYROLL_NODE_LABEL } from "../../../../../constants/payrollDeductions";
 
 const theme = createTheme();
 
@@ -56,13 +59,13 @@ describe("budgeting | chartColors", () => {
       [
         {
           categoryKey: "housing",
-          heading: "HOUSING",
+          heading: "Housing",
           total: 2000,
           color: "success",
           items: [],
         },
       ],
-      [{ name: "HOUSING", y: 2000 }],
+      [{ name: "Housing", y: 2000 }],
     );
 
     expect(data[0]?.color).toBe(theme.palette.success.main);
@@ -85,7 +88,7 @@ describe("budgeting | chartColors", () => {
     const data = colorizeIncomeOverviewPieData(theme, flow.categories, [
       { name: FEDERAL_TAX_LABEL, y: flow.federalTax },
       { name: PAYROLL_NODE_LABEL, y: flow.totalPayrollDeductions },
-      { name: "HOUSING", y: 24_000 },
+      { name: "Housing", y: 24_000 },
     ]);
 
     expect(data[0]?.color).toBe(theme.palette.error.main);
