+++ src/components/resume/finances/budgeting/graphs/chartData.test.ts
@@ -1,35 +1,33 @@
 import { describe, expect, it } from "vitest";
 import {
   buildBudgetFlow,
+  buildCategoryTotals,
+  getLatestBudgetIncome,
+  resolveExpenseAmount,
+  type ExpenseEntry,
+} from "../../../../../apis/budget";
+import {
   buildBudgetSankeyData,
   buildCategoryPieData,
-  buildCategoryTotals,
   buildExpensePieData,
   buildIncomeOverviewPieData,
   buildPayrollPieData,
-  computeCaliforniaTax,
-  computeFederalTax,
-  computeTotalTax,
-  formatPercentSources,
-  getLatestBudgetIncome,
-  getPercentSources,
+  getSankeyNodeClassName,
   getSankeyNodeSum,
   GROSS_INCOME_NODE,
   INCOME_NODE_LABELS,
   isCategorySankeyNode,
   isPayrollSankeyNode,
-  resolveExpenseAmount,
   UNALLOCATED_NODE,
-} from "./helpers";
-import { FEDERAL_TAX_LABEL } from "../../../../constants/federalTaxBrackets";
-import { STATE_TAX_LABEL } from "../../../../constants/caStateTaxBrackets";
+} from "./chartData";
+import { FEDERAL_TAX_LABEL } from "../../../../../constants/federalTaxBrackets";
+import { STATE_TAX_LABEL } from "../../../../../constants/caStateTaxBrackets";
 import {
   CA_DISABILITY_LABEL,
   MEDICARE_LABEL,
   PAYROLL_NODE_LABEL,
   SOCIAL_SECURITY_LABEL,
-} from "../../../../constants/payrollDeductions";
-import type { ExpenseEntry } from "../../../../jotai/finances-atom";
+} from "../../../../../constants/payrollDeductions";
 
 const sampleIncome = getLatestBudgetIncome(100_000, 20_000, 1_000, 2_000);
 
@@ -45,170 +43,25 @@ const sankeyNodeColors = {
   category: () => "#h",
 };
 
-describe("budgeting | helpers", () => {
-  describe("computeFederalTax", () => {
-    it("returns zero when gross is below the standard deduction", () => {
-      expect(computeFederalTax(10_000)).toBe(0);
-    });
+describe("resume | finances | budgeting | graphs | chartData", () => {
+  it("annualizes percent expenses in the sankey", () => {
+    const income = getLatestBudgetIncome(282_000, 50_000, 0, 32_281.78);
+    const entry: ExpenseEntry = {
+      name: "401k",
+      category: "Retirement",
+      value: 9,
+      valueMode: "percent",
+      percentSources: ["salary", "bonus"],
+    };
+    const monthly = resolveExpenseAmount(entry, income);
+    const flow = buildBudgetFlow(income, [entry]);
+    const { data } = buildBudgetSankeyData(flow, sankeyNodeColors);
 
-    it("computes tax at a mid-bracket income", () => {
-      const tax = computeFederalTax(100_000);
-      expect(tax).toBeGreaterThan(0);
-      expect(tax).toBeLessThan(100_000);
-    });
-  });
-
-  describe("computeCaliforniaTax", () => {
-    it("returns zero when gross is below the CA standard deduction", () => {
-      expect(computeCaliforniaTax(5_000)).toBe(0);
-    });
-
-    it("adds mental health surcharge above one million taxable", () => {
-      const belowSurcharge = computeCaliforniaTax(900_000);
-      const withSurcharge = computeCaliforniaTax(1_200_000);
-
-      expect(withSurcharge).toBeGreaterThan(belowSurcharge);
-    });
-  });
-
-  describe("computeTotalTax", () => {
-    it("sums federal and state tax", () => {
-      const { federal, state, total } = computeTotalTax(150_000);
-
-      expect(total).toBeCloseTo(federal + state, 2);
-    });
-  });
-
-  describe("resolveExpenseAmount", () => {
-    it("returns fixed dollar amounts", () => {
-      const entry: ExpenseEntry = {
-        name: "Rent",
-        category: "Housing",
-        value: 2500,
-      };
-
-      expect(resolveExpenseAmount(entry, sampleIncome)).toBe(2500);
-    });
-
-    it("resolves percent of salary", () => {
-      const entry: ExpenseEntry = {
-        name: "401k",
-        category: "Retirement",
-        value: 9,
-        valueMode: "percent",
-        percentSources: ["salary"],
-      };
-
-      expect(resolveExpenseAmount(entry, sampleIncome)).toBe(750);
-    });
-
-    it("resolves percent of stock adj", () => {
-      const entry: ExpenseEntry = {
-        name: "Invest",
-        category: "Investing",
-        value: 100,
-        valueMode: "percent",
-        percentSources: ["stockAdj"],
-      };
-
-      expect(resolveExpenseAmount(entry, sampleIncome)).toBeCloseTo(
-        2000 / 12,
-        2,
-      );
-    });
-
-    it("resolves percent across multiple income sources", () => {
-      const entry: ExpenseEntry = {
-        name: "401k",
-        category: "Retirement",
-        value: 9,
-        valueMode: "percent",
-        percentSources: ["salary", "bonus"],
-      };
-
-      expect(resolveExpenseAmount(entry, sampleIncome)).toBe(900);
-    });
-
-    it("annualizes to the correct annual percent in the sankey", () => {
-      const income = getLatestBudgetIncome(282_000, 50_000, 0, 32_281.78);
-      const entry: ExpenseEntry = {
-        name: "401k",
-        category: "Retirement",
-        value: 9,
-        valueMode: "percent",
-        percentSources: ["salary", "bonus"],
-      };
-      const monthly = resolveExpenseAmount(entry, income);
-      const flow = buildBudgetFlow(income, [entry]);
-      const { data } = buildBudgetSankeyData(flow, sankeyNodeColors);
-
-      expect(monthly).toBeCloseTo((0.09 * (282_000 + 50_000)) / 12, 2);
-      expect(data.find((link) => link.to === "RETIREMENT")?.weight).toBeCloseTo(
-        0.09 * (282_000 + 50_000),
-        2,
-      );
-    });
-  });
-
-  describe("buildCategoryTotals", () => {
-    it("groups expenses case-insensitively", () => {
-      const entries: ExpenseEntry[] = [
-        { name: "Rent", category: "housing", value: 2000 },
-        { name: "HOA", category: "Housing", value: 300 },
-      ];
-
-      const totals = buildCategoryTotals(entries, sampleIncome);
-
-      expect(totals).toHaveLength(1);
-      expect(totals[0]?.total).toBe(2300);
-      expect(totals[0]?.heading).toBe("HOUSING");
-    });
-
-    it("uses an uncategorized heading for blank categories", () => {
-      const totals = buildCategoryTotals(
-        [{ name: "Misc", category: "   ", value: 50 }],
-        sampleIncome,
-      );
-
-      expect(totals[0]?.heading).toBe("UNCATEGORIZED");
-    });
-  });
-
-  describe("buildBudgetFlow", () => {
-    it("computes unallocated net after withholdings and expenses", () => {
-      const flow = buildBudgetFlow(sampleIncome, [
-        { name: "Rent", category: "Housing", value: 3000 },
-      ]);
-
-      expect(flow.totalWithholdings).toBeCloseTo(
-        flow.totalTax + flow.totalPayrollDeductions,
-        2,
-      );
-      expect(flow.net).toBe(sampleIncome.gross - flow.totalWithholdings);
-      expect(flow.unallocated).toBeCloseTo(flow.net - 3000 * 12, 2);
-      expect(flow.isOverAllocated).toBe(false);
-    });
-
-    it("includes payroll deductions on gross wages", () => {
-      const flow = buildBudgetFlow(sampleIncome, []);
-
-      expect(flow.socialSecurity).toBeGreaterThan(0);
-      expect(flow.medicare).toBeGreaterThan(0);
-      expect(flow.caDisability).toBeGreaterThan(0);
-      expect(flow.totalPayrollDeductions).toBeCloseTo(
-        flow.socialSecurity + flow.medicare + flow.caDisability,
-        2,
-      );
-    });
-
-    it("flags over-allocation when expenses exceed net", () => {
-      const flow = buildBudgetFlow(sampleIncome, [
-        { name: "Everything", category: "Spend", value: 500_000 },
-      ]);
-
-      expect(flow.isOverAllocated).toBe(true);
-      expect(flow.unallocated).toBeLessThan(0);
-    });
+    expect(monthly).toBeCloseTo((0.09 * (282_000 + 50_000)) / 12, 2);
+    expect(data.find((link) => link.to === "Retirement")?.weight).toBeCloseTo(
+      0.09 * (282_000 + 50_000),
+      2,
+    );
   });
 
   describe("buildBudgetSankeyData", () => {
@@ -256,7 +109,7 @@ describe("budgeting | helpers", () => {
       ]);
       const { data } = buildBudgetSankeyData(flow, sankeyNodeColors);
 
-      expect(data.find((link) => link.to === "HOUSING")?.weight).toBe(36_000);
+      expect(data.find((link) => link.to === "Housing")?.weight).toBe(36_000);
     });
 
     it("includes unallocated when net exceeds allocated expenses", () => {
@@ -281,7 +134,7 @@ describe("budgeting | helpers", () => {
       ).toBe(false);
       expect(data.some((link) => link.to === FEDERAL_TAX_LABEL)).toBe(false);
       expect(data.some((link) => link.to === STATE_TAX_LABEL)).toBe(false);
-      expect(data.some((link) => link.to === "SPEND")).toBe(false);
+      expect(data.some((link) => link.to === "Spend")).toBe(false);
     });
 
     it("omits unallocated when expenses consume all net income", () => {
@@ -333,11 +186,11 @@ describe("budgeting | helpers", () => {
           FEDERAL_TAX_LABEL,
           STATE_TAX_LABEL,
           PAYROLL_NODE_LABEL,
-          "HOUSING",
+          "Housing",
           UNALLOCATED_NODE,
         ]),
       );
-      expect(pie.find(({ name }) => name === "HOUSING")?.y).toBe(24_000);
+      expect(pie.find(({ name }) => name === "Housing")?.y).toBe(24_000);
     });
 
     it("includes withholdings even when there are no expense categories", () => {
@@ -377,7 +230,7 @@ describe("budgeting | helpers", () => {
       const pie = buildIncomeOverviewPieData(flow);
 
       expect(pie.map(({ name }) => name)).not.toContain(UNALLOCATED_NODE);
-      expect(pie.some(({ name }) => name === "SPEND")).toBe(true);
+      expect(pie.some(({ name }) => name === "Spend")).toBe(true);
     });
   });
 
@@ -434,54 +287,6 @@ describe("budgeting | helpers", () => {
     });
   });
 
-  describe("getLatestBudgetIncome", () => {
-    it("prefers stockAdj over stock", () => {
-      expect(getLatestBudgetIncome(100, 0, 50, 75)).toEqual({
-        salary: 100,
-        bonus: 0,
-        stockAdj: 75,
-        gross: 175,
-      });
-    });
-
-    it("falls back to stock when stockAdj is zero", () => {
-      expect(getLatestBudgetIncome(100, 0, 50, 0)).toEqual({
-        salary: 100,
-        bonus: 0,
-        stockAdj: 50,
-        gross: 150,
-      });
-    });
-  });
-
-  describe("getPercentSources", () => {
-    it("prefers percentSources when present", () => {
-      expect(
-        getPercentSources({
-          percentSources: ["bonus", "stockAdj"],
-        }),
-      ).toEqual(["bonus", "stockAdj"]);
-    });
-
-    it("falls back to legacy percentSource", () => {
-      expect(getPercentSources({ percentSource: "stockAdj" })).toEqual([
-        "stockAdj",
-      ]);
-    });
-
-    it("defaults to salary when no sources are provided", () => {
-      expect(getPercentSources({})).toEqual(["salary"]);
-    });
-  });
-
-  describe("formatPercentSources", () => {
-    it("formats all supported income sources", () => {
-      expect(formatPercentSources(["salary", "bonus", "stockAdj"])).toBe(
-        "salary + bonus + stock adj",
-      );
-    });
-  });
-
   describe("isCategorySankeyNode", () => {
     it("matches sankey nodes by category heading", () => {
       const categories = buildCategoryTotals(
@@ -489,8 +294,40 @@ describe("budgeting | helpers", () => {
         sampleIncome,
       );
 
-      expect(isCategorySankeyNode("HOUSING", categories)).toBe(true);
+      expect(isCategorySankeyNode("Housing", categories)).toBe(true);
       expect(isCategorySankeyNode("Fed Tax", categories)).toBe(false);
+    });
+  });
+
+  describe("getSankeyNodeClassName", () => {
+    const categories = buildCategoryTotals(
+      [{ name: "Rent", category: "Housing", value: 2000 }],
+      sampleIncome,
+    );
+
+    it("returns selected class for a matching category node", () => {
+      expect(getSankeyNodeClassName("Housing", categories, "housing")).toBe(
+        "budget-sankey-selected",
+      );
+    });
+
+    it("returns selected class for payroll when payroll is selected", () => {
+      expect(
+        getSankeyNodeClassName(PAYROLL_NODE_LABEL, categories, "payroll"),
+      ).toBe("budget-sankey-selected");
+    });
+
+    it("returns undefined when nothing is selected", () => {
+      expect(
+        getSankeyNodeClassName("Housing", categories, null),
+      ).toBeUndefined();
+      expect(getSankeyNodeClassName("Housing", categories)).toBeUndefined();
+    });
+
+    it("returns undefined when another category is selected", () => {
+      expect(
+        getSankeyNodeClassName("Housing", categories, "food"),
+      ).toBeUndefined();
     });
   });
 });
